Python-EDA Portfolio - Treadmill Buyer Profile
The market research team at AeroFit wants to identify the characteristics of the target 
audience for each type of treadmill offered by the company, to provide a better 
recommendation of the treadmills to new customers. The team decides to investigate 
whether there are differences across the product with respect to customer 
characteristics.